Refactor to create senders together for code reuse
