+++ app/index.js
@@ -5,6 +5,10 @@ const init = async () => {
 
   await server.start()
   console.log(`Server running on ${server.info.uri}`)
+
+  const senders = require('./messaging/queue-senders')
+  senders.updateAgreement({ test: 'new test from frontend' })
+  senders.updateEligibility({ test: 'new test from frontend' })
 }
 
 process.on('unhandledRejection', (err) => {
